pyproject.toml
[tool.poetry]
name = "cs439-miniproject"
version = "0.1.0"
description = ""
authors = ["charlesbvll <charles.beauville@epfl.ch>"]
readme = "README.md"
packages = [{include = "cs439_miniproject/src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
tqdm = "^4.65.0"
torchvision = "^0.15.1"
flwr = "^1.3.0"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
ray = "^2.3.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
